backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
supabase==2.3.0
google-generativeai==0.3.2
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0 
dj-database-url 